+++ CHANGELOG.md
@@ -19,11 +19,12 @@ The minor version will be incremented upon a breaking change and the patch versi
 ## [11.2.0]
 
 ### Features
-- Added support to block transaction signed by a banned account. 
+- Added support to block transactions signed by a banned identity. 
 - Created file `blocking_services.rs` to store all logic related to blocking
 
 ### Fixes
-- Added tests for `BannedAccounts`
+- Refactor blocking leaders.
+- Added tests for `IdentitiesBlocker`
 
 ## [11.1.2]